+++ R/aaa.R
@@ -44,3 +44,5 @@ check_object <- function(x,
     call = call
   )
 }
+
+is_waiver <- function(x) inherits(x, "waiver")